requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Faker==19.3.0
drf-yasg==1.21.7 